move entity download button logic to separate  component
